+++ src/schema/index.ts
@@ -9,6 +9,7 @@ import merge from "lodash/merge"
 import * as _typeDefs from "./typeDefs"
 import { AuthDirective, DevelopmentDirective } from "../directives"
 import { authResolvers } from "../modules/Auth"
+import { discussionResolvers } from "../modules/Discussion"
 import { s3Resolvers } from "../modules/S3"
 import { userResolvers } from "../modules/User"
 
@@ -17,6 +18,7 @@ const typeDefs: ITypeDefinitions = [...Object.values(_typeDefs)]
 const resolvers: IResolvers = merge(
   { Query: {}, Mutation: {} },
   authResolvers,
+  discussionResolvers,
   s3Resolvers,
   userResolvers
 )