+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.7.0] - 2020-12-14
+
+This version encompasses work for a POC of recreating GitHub's 'Discussions' feature
+
+### Added
+
+- installed `yup` to validate requests that pass GraphQL schema validation, but don't actually have valid values
+- `Query.getDiscussionById`
+- `Query.getDiscussions`
+- `Mutation.createDiscussion`
+- `Mutation.createComment`
+
+#### DynamoDB `GSI`'s
+
+- **`GSI_SK_Created`**
+  - Partition: `'SK'`; Sort: `'created'`
+- **`GSI_Root_Created`**
+  - Partition: `'Root'`; Sort: `'created'`
+
+### Updated
+
+- types.d.ts
+- schema
+
+## [v0.6.3] - 2020-12-10
+
+### Added
+
+- `uuid` package
+- `getSelectedFields` util
+
+### Fixed
+
+- `getOrCreateUser` saves `id`
+
 ## [v0.6.2] - 2020-12-06
 
 ### Changed
@@ -240,6 +275,8 @@ add skip/take to getCommentsByUrl
 
 ### Added
 
+[v0.7.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.6.3...v0.7.0
+[v0.6.3]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.6.2...v0.6.3
 [v0.6.2]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.6.1...v0.6.2
 [v0.6.1]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.6.0...v0.6.1
 [v0.6.0]: https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.5.1...v0.6.0